Moves DigitalOceanCreateServer to its own directory
